用户列表增删改查

+++ ss-service/src/main/java/com/rainbow/service/ISysDeptService.java
@@ -15,7 +15,7 @@ public interface ISysDeptService {
 
     void addDept(SysDeptReq deptReq);
 
-    void deleteDept(int deptId);
+    void deleteDept(Long deptId);
 
     void updateDept(SysDeptReq deptReq);
